chore : 홈페이지와 메인페이지를 분리
